pyproject.toml
[tool.poetry]
name = "newsbot"
version = "0.1.0"
description = ""
authors = ["Ra1ze505 <87398772+Ra1ze505@users.noreply.github.com>"]

[tool.poetry.dependencies]
python = "^3.10"
SQLAlchemy = "^1.4.36"
alembic = "^1.7.7"
Telethon = "^1.24.0"
psycopg2 = "^2.9.3"
asyncpg = "^0.25.0"
greenlet = "^1.1.2"
aiohttp = "^3.8.1"
celery = "^5.2.6"
redis = "^4.3.1"
db = "^0.1.1"

[tool.poetry.dev-dependencies]
python-dotenv = "^0.20.0"
pytest = "^7.1.2"
pytest-asyncio = "^0.18.3"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
minversion = "6.0"
addopts = "-ra -q"
testpaths = [
    "tests",
]
pythonpath = [
  ".", "src/",
]
asyncio_mode = 'auto'